// src/resources/Register_userUserInterface.ts
export enum Register_userUserInterface {

    buttom_account = '//*[@id="top-links"]/ul/li[2]/a',
    buttom_register = '//*[@id="top-links"]/ul/li[2]/ul/li[1]/a',
    input_first_name = '//*[@id="input-firstname"]',
    input_last_name = '//*[@id="input-lastname"]',
    input_e_mail = '//*[@id="input-email"]',
    input_telephone = '//*[@id="input-telephone"]',
    input_password = '//*[@id="input-password"]',
    input_password_confirm = '//*[@id="input-confirm"]',
    check_policy = '//*[@id="content"]/form/div/div/input[1]',
    buttom_continue = '//*[@id="content"]/form/div/div/input[2]',
    buttom_continue2 = '//*[@id="content"]/div/div/a'

}